+++ spec/internal/db/migrate/20260430151510_create_ductwork_executions.rb
@@ -13,6 +13,7 @@ class CreateDuctworkExecutions < Ductwork::Migration
       table.timestamp :started_at, null: false
       table.timestamp :completed_at
       table.integer :retry_count, null: false
+      table.integer :crash_count, null: false
       table.timestamps null: false
     end